uyghurocr-api/requirements.txt
fastapi
uvicorn[standard]
onnxruntime
opencv-python
numpy
pytesseract
pymupdf
pyclipper
python-multipart
pillow
